refactor(radiant): centralize legacy host readiness
